+++ server/api/src/main/kotlin/com/kw/api/domain/member/dto/request/BundleOrderUpdateRequest.kt
@@ -1,0 +1,3 @@
+package com.kw.api.domain.member.dto.request
+
+data class BundleOrderUpdateRequest(val bundleIds: List<Long>)
