--- server/api/src/main/kotlin/com/kw/api/domain/member/dto/request/MemberBundleOrderUpdateRequest.kt
@@ -1,3 +1,0 @@
-package com.kw.api.domain.member.dto.request
-
-data class MemberBundleOrderUpdateRequest(val bundleIds: List<Long>)
